requirements.txt
flask==3.0.0
python-dotenv==1.0.0
groq==1.0.0
pypdf==4.0.1
truststore==0.9.1
werkzeug==3.0.1
reportlab==4.0.4
